Use upstream k-s-o in 1.4.0 CSV
This involves referring to a new image so new env needed. Also,
because we use $ in some of the values, e.g. ${NAME}, we want to be
particular about exactly what gets substituted.

Minor changes to README to reflect the fact that this is only going to
work on recent versions of OCP, so complex instructions to support
minikube are useless.