Upgrade MIN_OPENSHIFT_VERSION to 4.3.0-0.nightly.
KUBERNETES_MIN_VERSION is v1.15.0 for the upstream eventing operator built from master, which comes with OCP 4.3. Actually OCP 4.3 comes with Kube 1.16 and 4.2 is 1.14.

We need to use the upstream eventing operator from master because 0.11 of it uses a different CRD kind and I don't want to support that case.