Use upstream v0.11.1 knative-serving operator (#66)
* Use upstream k-s-o in 1.4.0 CSV

This involves referring to a new image so new env needed. Also,
because we use $ in some of the values, e.g. ${NAME}, we want to be
particular about exactly what gets substituted.

Minor changes to README to reflect the fact that this is only going to
work on recent versions of OCP, so complex instructions to support
minikube are useless.

* Update test infra to reflect CRD group name change

Also enable sequential test runs, with the occasional `make teardown`

Handy to delete all CSV's for upgrade tests since multiple may be installed.

The -x is too noisy; exporting LOG_LEVEL=DEBUG may be better

* Update deps for the v0.11.1 upstream k-s-o

* Orphan old CR's resources to avoid potential dropped requests

* Account for the different cert configs between controllers

* Introduce top-level KNATIVE_VERSION var

Eventually, we may want to default this from the CSV or the
Gopkg.toml, but until the structure of the tag is settled, we'll have
to manually set this each time we bump the version