Support both the old and the new group of KnativeServings. (#91)
This adds all the changes necessary to support knativeservings.serving.knative.dev and knativeservings.operator.knative.dev in one version. The old (serving) based API is supported by creating an entity of the new resource as a "child" of the old one. The `Spec` field of the old resource will be mirrored "down" into the new one, any status changes on the new one will be reflected in the old resource. That should make the old resource behave exactly like before, enabling this move to be backwards compatible.
Before creating the new resource, all of the old resources need to be orphaned to enable the new KnativeServing to gain ownership over all of these resources.
Added testcases ensure that different permutations of KnativeServings work as expected.

The migration path now looks like this:
1. Install the new version of the operators.
2.a. An installed system keeps working as before. You will have both an old and a new resource.
2.b. You can install the system freshly with the old resource and it still works as before. You will have both an old and a new resource.
2.c. You can install the system with the new CR. You will have only that new resource.
3. If you had both old and a new resource, you now have time to adjust your scripts etc. Once you've done that, you can remove the old resource without repercussions and continue to use just the new resource.

**Note:** If you want to adjust the Spec via the new CR, you **must** first delete the old CR.

Co-authored-by: Jim Crossley <jim@crossleys.org>
Co-authored-by: Martin Gencur <mgencur@redhat.com>

Co-authored-by: Martin Gencur <mgencur@redhat.com>
Co-authored-by: Jim Crossley <jim@crossleys.org>