[SRVKS-427] Override all configuration we set as we have authority. Apply mutations on update.
Currently, when we update the operator, we will try to override the image setting in the respective config maps. Our `Configure` helper is too conservative though, and only writes config that was not set before.

All configs that we actively set are subject to our authority though. For now, we should always override all of these settings as they are integral to our integration.