Overwrite config after upgrade [SRVKS-427]
The mutation annotation is now the operator's version instead of a
timestamp.

Configure always overwrites, but Mutate will only run its stages if
the annotated version is different than its own.

There is a redundant check for the annotation in the controller to
avoid updating the ks status, but the 2nd check should rarely be
necessary -- theoretically once per upgrade.

This seems simple and good, so I'm probably forgetting something
obvious :)