Use a differently named service to workaround kourier bug.
We're hunting a bug in Kourier, which apparently causes issues if we quickly delete and recreate the same service. As the general e2e/conformance test suite seems not affected and as we do have a reliable reproducer, we're going to workaround the issues in our CI for now to unblock other work.