[SRVKS-513] Manually combine configmaps for service-ca and trusted-ca.
My assumption about being able to use both the serviceCA annotation and the trustedCA label on the same configMap have been proven wrong. This pulls their management apart into separate configMaps that are watched by the reconciler. The contents of both configMaps are unified into a single configMap, which is then mounted into the controller via the operator's mechanism.