Limit component CR's to 1 per cluster
This is especially critical for upgrades, in which say, a 1.8.0
install has resources in the knative-serving namespace, but the 1.9.0
operator only allows new ones in openshift-serverless. Without this
change, that logic could result in 2 CR's installed on the cluster.

TODO: refactor these validations to take a KComponent interface
instead of the specific subtypes so that they can share code.