Move to the new bundle format for 1.7.3
This replaces the "old" way of doing OLM manifests with the new bundle format. (#409)

Approach

The way our CatalogSource is built has been vastly revamped. Instead of relying on a ConfigMap that hosts a list of CSVs, we now rely on a gRPC based source. During the run, a new bundle image (containing the CSV and metadata for the **current** version) will be built in the cluster-internal registry of the cluster under test. We'll then launch a `serverless-index` deployment, which contains the gRPC registry, which will pull the respective bundle image from the internal registry. In addition to that, it's also fed a bundle image for the version we want to upgrade from (1.7.2 in this case). This has been built [manually for this version](https://github.com/openshift-knative/serverless-operator/pull/403), seeing as 1.7.2 is still on the old format. For 1.9.0 onwards, we'll be able to use the promoted images by CI as 1.8 will be building a serverless-bundle as part of its images.

Notable changes

1. We now have to define which is the current and the next CSV. Especially the former can likely be inferred from the files themselves. I'd like to leave that to a followup PR.
2. The in-repository CSV is now a "working" CSV by default. It contains the to-be-promoted images so that the image built by CI (which won't replace any image coordinates) has the right image coordinates built into it.

Co-authored-by: Markus Thömmes <markusthoemmes@me.com>
Co-authored-by: Ben Browning <bbrownin@redhat.com>