[RELEASE-1.7] Move to the new bundle format (#414)
* Move to the new bundle format for 1.7.3

This replaces the "old" way of doing OLM manifests with the new bundle format. (#409)

Approach

The way our CatalogSource is built has been vastly revamped. Instead of relying on a ConfigMap that hosts a list of CSVs, we now rely on a gRPC based source. During the run, a new bundle image (containing the CSV and metadata for the **current** version) will be built in the cluster-internal registry of the cluster under test. We'll then launch a `serverless-index` deployment, which contains the gRPC registry, which will pull the respective bundle image from the internal registry. In addition to that, it's also fed a bundle image for the version we want to upgrade from (1.7.2 in this case). This has been built [manually for this version](https://github.com/openshift-knative/serverless-operator/pull/403), seeing as 1.7.2 is still on the old format. For 1.9.0 onwards, we'll be able to use the promoted images by CI as 1.8 will be building a serverless-bundle as part of its images.

Notable changes

1. We now have to define which is the current and the next CSV. Especially the former can likely be inferred from the files themselves. I'd like to leave that to a followup PR.
2. The in-repository CSV is now a "working" CSV by default. It contains the to-be-promoted images so that the image built by CI (which won't replace any image coordinates) has the right image coordinates built into it.

Co-authored-by: Markus Thömmes <markusthoemmes@me.com>
Co-authored-by: Ben Browning <bbrownin@redhat.com>

* Swap current and previous csv values

* Reduce hacks needed for bundle format CI. (#412)

- Dropped the need to run with a hacked version of OPM by specifying the internal registry as an insecure registry to make podman do "the right thing" automatically.
- Dropped the need to run the index image as a root by specifying the needed access in the image.
- Checked in the Dockerfile to produce the image here to make rebuilding easier if ever needed.

* Revert some changes made in e4b6518

* Add function add_networkpolicy to test lib

* Fix bug with sed command

* Fix One Click and use --kubeconfig

* Update image references

* Use 1.7.2 bundle image for 4.3,4.4,4.5 channels

And other small changes, including bumping proxyv2 image since
the readiness probes were failing with proxyv2:1.0.8.

* Use proxyv2-ubi8:1.0.8, same as in 1.7.2

* Use latest opm binary, replace nightly with v1.7.3

* Try fixing missing images

* Use the correct variable

* Remove unused add_networkpolicy function

* Update TODO comment

* Revert changes to test/serving.bash with minor mod

* Remove hardcoded registry host using IMAGE_FORMAT

* Fix error seen during local builds

* Remove 4.5 channel

* Use add-scc-to-user privileged

* Add securityContext privileged true on index pod

* Fix undefined var during local upgrade test

* Bump to serverless-bundle:1.7.2-1

* Bump SCALE_UP from 6 to 8

* Bump SCALE_UP from 8 to 10

* Read current and previous version from CSV

* Remove 4.3 channel and bump default channel to 4.4

* Remove go get github.com/mikefarah/yq/v3

* Define latest_serving_version

In addition, remove 4.1 and 4.2 specific code as we no longer use
these versions of OCP in testing.

* Remove prev_serving_version

* Update references to 4.4 docs

Co-authored-by: Markus Thömmes <markusthoemmes@me.com>
Co-authored-by: Ben Browning <bbrownin@redhat.com>