[SRVKS-590] Don't reconcile KS/KE if placed in the wrong namespace.
Apart from the checks in the webhooks, this aborts reconcilation for
Knative Serving/Knative Eventing resources if they don't have been
placed into the expected namespace. This can happen if the resources are
placed before the webhook is up and similar cases. An error message is
surfaced in the status of the resource.