[SRVKS-741] Drop namespaces from manifests to prevent reconcilation
Having the namespaces in the manifests doesn't do anything useful as
we're requiring the user to create the namespace beforehand anyway. It
does however actively cause us to drop labels and annotations that users
have added via `kubectl apply` commands (versus `kubectl annotate` or
`kubectl label`), so it's actually harmful.