Explicitly set controller deployment name to override controller image
Upstream changed Kourier controller's container name to controller.

It is duplicated with Serving controller and so we need to explicitly
set the `<deploy_name>`/`<container_name>` to override the image.

The net-kourier deployment name is not changed in the downstream so
this PR changes only serving controller.
(And it works with nigthly test in https://github.com/openshift/knative-serving)