Add metadata-webhook code
This patch adds the code https://github.com/nak3/metadata-webhook into
serverless-operator/experimental.

The metadata-webhook is used for the mesh e2e test to add annotation.
The code was hosted on my personal repo but it should be hosted in
serverless-operator repo as it needs to keep pkg versions consistent.

Also, it adds github action to build metadata-webhook image and push
to ghcr.io.