Remove redundant ContinualTest from knative/eventing
The test is identical to BrokerBackedByChannelTest because in
serverless-operator repo we set up KafkaChannel as the default channel
before the test suite. Both tests deploy Broker and Trigger and use
KafkaChannel in the background. The tests are vendored from different
repositories and due to the settings we have they test the same
configuration.