Change sample-rate for tracing in upgrade tests to 0.3 (#1611)
* Change sample-rate for tracing in upgrade tests to 0.3

This is to prevent OOM issues in Zipkin.

Having sample-rate 0.1 produces about 2,106.8 MiB of data in Zipkin. We
currently have -Xmx9G for Zipkin so setting this parameter to 0.3 in
order to have some buffer for cases when the tests take longer and
produce more data.

* sample-rate 0.3 only for upgrade tests

For E2E tests, use the default 1.0