Improve isStatefulSetAvailable condition check
On upgrades and during normal operation the number of replicas might vary, however, we
shouldn't transition to "not ready" just because there are some non-ready replicas because that
could  trigger a warning for anyone using KnativeKafka that it's not a signal of a
non-functional installation.

Note: we do this only in the case where the number of expected replicas is >= 1, so that we
make sure that at least one replica is ready and functional.

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>