Reduce trace export limit from default 100 to 30 for each test
The goal is to speed up test execution as getting traces for many events
takes a lot of time.
There are two types of failures: missed events and duplicate events.
Usually, when it happens the number of missed events is very low (<10)
so this limit is for duplicate events. We allow duplicate events in the
test and traces are only informational.