Image template for Serving points to multiarch images (#2071)
The path for publishing Serving images was recently changed in
openshift-knative/serving repo. The template was updated to point to the
new images. However, there are also Github actions in
openshift-knative/serving which still build "multiarch" images and push
them to quay and serverless-operator needs to use those images so as to
allow testing on P/Z.

Co-authored-by: Martin Gencur <mgencur@redhat.com>