Automation for knative-istio-authz-chart bump
This introduces a new automation for the bump of the
knative-istio-authz-chart helm chart metadata.

It will bump the chart metadata in `Chart.yaml` when the
same SO bump automation is triggered.

In addition, since SO bump itself depends on the chart
availability, it temporarly configures the `make generated-files`
command to use the "current SO version" as opposed to the
"bumped/next version" so that we don't need to retry until
it's available.

Signed-off-by: Pierangelo Di Pilato <pierdipi@redhat.com>