Fix upgrade sequence for kitchensink tests (#2277)
* Fix upgrade sequence for kitchensink tests

* 1.30.0 now replaces 1.29.0 so 1.29.1 is not offered for upgrades
* 1.30.0 was released so it's not in redhat-operators Catalog Sourec

* Increase SCALE_UP to 4

* Try redhat-operator and then serverless-operator catalog source

In kitchensink tests. This approach doesn't require specifying
catalogsource name in project.yaml for each version.

* Use redhat-operators cs for kitchensink upgrade tests for initial install

* Fallback to trying serverless-operator cs only if the previous one was different

* Separate post-upgrade and post-downgrade tests

All tests within one phase run in parallel so it might happen that the
namespace is being terminated while we still want to preform some
verifications with the namespace (namespace deletion is done in
post-downgrade tests)

* Check error against wait.ErrWaitTimeout