fix(subscription): Fix ambiguous shortName
`subscriptions.operators.coreos.com` and `subscriptions.messaging.knative.dev` both have shortName `sub` that is causing confusion for the client (`oc`). While using FQNs should be the best approach, there are documentations like this
https://docs.openshift.com/container-platform/4.13/support/troubleshooting/troubleshooting-operator-issues.html#olm-status-viewing-cli_troubleshooting-operator-issues across openshift and there's a chance there may be scripts relying on this. This is not a new issue
and there's an outstanding debate as to whether this is a bug and how the client should handle such ambiguity, nothing is looking like a quick solution.

However, this is causing immediate problems when knative and coreos are being used in a cluster. Due to the way how the version priority works wrt clients (Refer
https://kubernetes.io/docs/tasks/extend-kubernetes/custom-resources/custom-resource-definition-versioning/#version-priority), with the versions being the same, there's no guarantee which sub is returned when commands like `oc get sub` is called.

Since I'm not sure of the level of impact this has to the documentation across openshift, this PR may be a quick fix. I've opened a ticket for fixing the documentation but I'll let your team to decide taking this PR or close in favor of doc change.

 # Please enter
the commit message
for your changes. Lines starting # with '#' will be ignored, and an empty message aborts the commit.