Use v2 endpoint for getting alerts (#2770)
* Use v2 endpoint for getting alerts

Calling v1 endpoint on OCP 4.17 returns this:
{"status":"deprecated","error":"The Alertmanager v1 API was deprecated
in version 0.16.0 and is removed as of version 0.28.0 - please use the
equivalent route in the v2 API"}

The output of the v2 endpoint is an array of events while the v1
endpoint was a json object with .data element that included the array of
events. Thus chanding the jq query a bit.

* Select relevant alerts related to APIRemoved and compare with all allerts